requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
stupidArtnet>=1.3.0
websockets>=12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
jinja2>=3.1.0
